// src/app/Rest/Conductor/model/task.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Any } from './any';
import { TaskDef } from './taskDef';
import { WorkflowTask } from './workflowTask';


export interface Task {
    taskType?: string;
    status?: Task.StatusEnum;
    inputData?: { [key: string]: any; };
    referenceTaskName?: string;
    retryCount?: number;
    seq?: number;
    correlationId?: string;
    pollCount?: number;
    taskDefName?: string;
    scheduledTime?: number;
    startTime?: number;
    endTime?: number;
    updateTime?: number;
    startDelayInSeconds?: number;
    retriedTaskId?: string;
    retried?: boolean;
    executed?: boolean;
    callbackFromWorker?: boolean;
    responseTimeoutSeconds?: number;
    workflowInstanceId?: string;
    workflowType?: string;
    taskId?: string;
    reasonForIncompletion?: string;
    callbackAfterSeconds?: number;
    workerId?: string;
    outputData?: { [key: string]: any; };
    workflowTask?: WorkflowTask;
    domain?: string;
    inputMessage?: Any;
    outputMessage?: Any;
    rateLimitPerFrequency?: number;
    rateLimitFrequencyInSeconds?: number;
    externalInputPayloadStoragePath?: string;
    externalOutputPayloadStoragePath?: string;
    workflowPriority?: number;
    executionNameSpace?: string;
    isolationGroupId?: string;
    iteration?: number;
    subWorkflowId?: string;
    taskDefinition?: TaskDef;
    taskStatus?: Task.TaskStatusEnum;
    queueWaitTime?: number;
    loopOverTask?: boolean;
}
export namespace Task {
    export type StatusEnum = 'IN_PROGRESS' | 'CANCELED' | 'FAILED' | 'FAILED_WITH_TERMINAL_ERROR' | 'COMPLETED' | 'COMPLETED_WITH_ERRORS' | 'SCHEDULED' | 'TIMED_OUT' | 'SKIPPED';
    export const StatusEnum = {
        INPROGRESS: 'IN_PROGRESS' as StatusEnum,
        CANCELED: 'CANCELED' as StatusEnum,
        FAILED: 'FAILED' as StatusEnum,
        FAILEDWITHTERMINALERROR: 'FAILED_WITH_TERMINAL_ERROR' as StatusEnum,
        COMPLETED: 'COMPLETED' as StatusEnum,
        COMPLETEDWITHERRORS: 'COMPLETED_WITH_ERRORS' as StatusEnum,
        SCHEDULED: 'SCHEDULED' as StatusEnum,
        TIMEDOUT: 'TIMED_OUT' as StatusEnum,
        SKIPPED: 'SKIPPED' as StatusEnum
    }
    export type TaskStatusEnum = 'IN_PROGRESS' | 'CANCELED' | 'FAILED' | 'FAILED_WITH_TERMINAL_ERROR' | 'COMPLETED' | 'COMPLETED_WITH_ERRORS' | 'SCHEDULED' | 'TIMED_OUT' | 'SKIPPED';
    export const TaskStatusEnum = {
        INPROGRESS: 'IN_PROGRESS' as TaskStatusEnum,
        CANCELED: 'CANCELED' as TaskStatusEnum,
        FAILED: 'FAILED' as TaskStatusEnum,
        FAILEDWITHTERMINALERROR: 'FAILED_WITH_TERMINAL_ERROR' as TaskStatusEnum,
        COMPLETED: 'COMPLETED' as TaskStatusEnum,
        COMPLETEDWITHERRORS: 'COMPLETED_WITH_ERRORS' as TaskStatusEnum,
        SCHEDULED: 'SCHEDULED' as TaskStatusEnum,
        TIMEDOUT: 'TIMED_OUT' as TaskStatusEnum,
        SKIPPED: 'SKIPPED' as TaskStatusEnum
    }
}